Moved internal/cql_promise.hpp to cql_promise.hpp
