+++ include/cql/internal/cql_connection_impl.hpp
@@ -76,8 +76,8 @@
 #include "cql/cql_uuid.hpp"
 #include "cql/cql_stream.hpp"
 #include "cql/cql_host.hpp"
+#include "cql/cql_promise.hpp"
 #include "cql/internal/cql_util.hpp"
-#include "cql/internal/cql_promise.hpp"
 #include "cql/internal/cql_session_impl.hpp"
 #include "cql/exceptions/cql_query_timeout_exception.hpp"
 #include "cql/exceptions/cql_unavailable_exception.hpp"
